% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexing.R
\name{batch_names}
\alias{batch_names}
\title{Name the batch}
\usage{
batch_names(i)
}
\arguments{
\item{i}{A list of numeric indices.}
}
\value{
A character vector with the names of the batch for each the index.
}
\description{
Given an index return the name of the batches the samples are in
}
\examples{
index <- create_subset(100, 50, 2)
batch <- batch_names(index)
head(batch)
}
\seealso{
\code{\link[=create_subset]{create_subset()}}, for the inverse look at
\code{\link[=use_index]{use_index()}}.
}
