% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropy.R
\name{entropy}
\alias{entropy}
\title{Calculates the entropy}
\usage{
entropy(x)
}
\arguments{
\item{x}{A character or vector with two or more categories}
}
\value{
The numeric value of the Shannon entropy
}
\description{
Calculates the entropy of a category. It uses the amount of categories to
scale between 0 and 1.
}
\examples{
entropy(c("H", "T", "H", "T"))
}
