% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/own_algorithm.R
\name{evaluate_cat}
\alias{evaluate_cat}
\title{Evaluate category}
\usage{
evaluate_cat(i, pheno)
}
\arguments{
\item{i}{list of numeric indices of the data.frame}

\item{pheno}{Data.frame}
}
\value{
Value to minimize
}
\description{
Evaluate category
}
