% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_category.R
\name{evaluate_category}
\alias{evaluate_category}
\title{Evaluate category}
\usage{
evaluate_category(i, pheno)
}
\arguments{
\item{i}{list of numeric indices of the data.frame}

\item{pheno}{Data.frame with information about the samples}
}
\value{
Value to minimize
}
\description{
Looks if the nominal or character columns are equally distributed according
to the entropy and taking into account the independence between batches
}
\seealso{
\code{\link{evaluate_number}}, \code{\link{evaluate_independence}}
}
