% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/own_algorithm.R
\name{evaluate_num}
\alias{evaluate_num}
\title{Evaluate numbers}
\usage{
evaluate_num(i, pheno)
}
\arguments{
\item{i}{list of numeric indices of the data.frame}

\item{pheno}{Data.frame}
}
\value{
Value to minimize
}
\description{
Evaluate numbers
}
