% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_num.R
\name{evaluate_numbers}
\alias{evaluate_numbers}
\title{Evaluate numbers}
\usage{
evaluate_numbers(i, pheno)
}
\arguments{
\item{i}{list of numeric indices of the data.frame}

\item{pheno}{Data.frame}
}
\value{
Value to minimize, summarizing the difference with the desired output
}
\description{
Evaluate how far are each subset of the desired goal.
}
