% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC.R
\name{optimum_size}
\alias{optimum_size}
\title{Seek optimum size for a batch}
\usage{
optimum_size(pheno)
}
\arguments{
\item{pheno}{A \code{data.frame} with the information about the samples.}
}
\value{
A number for the number of samples needed.
}
\description{
Calculates the number of samples to representative of most of the variance
of their categorical variables.
}
