% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{position_name}
\alias{position_name}
\title{Create position names for a grid.}
\usage{
position_name(rows, columns)
}
\arguments{
\item{rows}{Names of the rows.}

\item{columns}{Names of the columns.}
}
\value{
A data.frame with the rows and columns and the resulting name row+column.
The name column is a factor for easier sorting in row, column order.
}
\description{
Create position names for a grid.
}
\examples{
position_name(c("A", "B"), 1:2)
}
