% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexing.R
\name{use_index}
\alias{use_index}
\title{Convert a factor to index}
\usage{
use_index(x)
}
\arguments{
\item{x}{A character or a factor to be used as index}
}
\description{
Convert a given factor to an accepted index
}
\examples{
plates <- c("P1", "P2", "P1", "P2", "P2", "P3", "P1", "P3", "P1", "P1")
use_index(plates)
}
\seealso{
You can use \code{\link[=evaluate_index]{evaluate_index()}} to evaluate how good an
index is. For the inverse look at  \code{\link[=batch_names]{batch_names()}}.
}
