% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexing.R
\name{batch_names}
\alias{batch_names}
\title{Name the batch}
\usage{
batch_names(i)
}
\arguments{
\item{i}{A list of numeric indices of the data}
}
\value{
A character vector with the names of the batch for each sample
}
\description{
Given an index return the name of the batches the samples are in
}
