% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC.R
\name{check_index}
\alias{check_index}
\title{Check index}
\usage{
check_index(pheno, index, omit = NULL)
}
\arguments{
\item{pheno}{Data.frame with the sample information.}

\item{index}{A list of indices indicating which samples go to which subset}

\item{omit}{Name of the columns of the pheno that will be omitted}
}
\value{
A matrix with the differences with the original data
}
\description{
Given an index
}
