% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/own_algorithm.R
\name{create_subset}
\alias{create_subset}
\title{Create index of subsets of a data}
\usage{
create_subset(size.subset, n, size.data)
}
\arguments{
\item{size.subset}{A numeric value with the amount of samples per batch}

\item{n}{A numeric value with the number of batchs}

\item{size.data}{A numeric value of the amount of samples to distribute}
}
\value{
A random list of indices of the samples
}
\description{
Create index of subsets of a data
}
\examples{
create_subset(50, 2, 100)
}
