% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/designer.R
\name{design}
\alias{design}
\title{Design a batch experiment}
\usage{
design(pheno, size.batch, omit = NULL, iterations = 500)
}
\arguments{
\item{pheno}{Data.frame with the sample information.}

\item{size.batch}{Numeric value of the number of sample per batch}

\item{omit}{Name of the columns of the pheno that will be omitted}

\item{iterations}{Numeric value of iterations that will be performed}
}
\value{
The indices of which samples go with which batch
}
\description{
Given some samples it distribute them in several batches, trying to have
equal number of samples per batch. It can handle both numeric and
categorical data.
}
