% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.R
\name{distribution}
\alias{distribution}
\title{Checks if all the values are maximally distributed in the several batches}
\usage{
distribution(report, column)
}
\arguments{
\item{report}{A data.frame which must contain a batch column. Which can be
obtained with \code{\link{insepct}}}

\item{column}{The name of the column one wants to inspect}
}
\value{
\code{TRUE} if the values are maximal distributed, otherwise
\code{FALSE}
}
\description{
Checks if all the values are maximally distributed in the several batches
}
