% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_category.R
\name{evaluate_independence}
\alias{evaluate_independence}
\title{Compare independence by chisq.test}
\usage{
evaluate_independence(i, pheno)
}
\arguments{
\item{i}{Index of subsets}

\item{pheno}{A data.frame with the information about the samples}
}
\value{
Returns a vector with the p-values of the chisq.test between the
category and the subset
}
\description{
Looks the independence between the categories and the batches
}
\seealso{
\code{\link{evaluate_category}}
}
