% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_num.R
\name{evaluate_mean}
\alias{evaluate_mean}
\title{Evaluates the mean of the numeric values}
\usage{
evaluate_mean(i, pheno)
}
\arguments{
\item{i}{List of indices}

\item{pheno}{Data.frame with information about the samples}
}
\value{
A matrix with the mean value for each column for each subset
}
\description{
Looks for the mean of the numeric values
}
