% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/own_algorithm.R
\name{evaluate_mean}
\alias{evaluate_mean}
\title{Evaluates the mean of the numeric values}
\usage{
evaluate_mean(i, pheno)
}
\arguments{
\item{i}{List of indices}

\item{pheno}{Data.frame with the samples}
}
\value{
A Matrix with the mean value for each column
}
\description{
Evaluates the mean of the numeric values
}
