% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropy.R
\name{evaluate_na}
\alias{evaluate_na}
\title{Evaluate how many values are empty}
\usage{
evaluate_na(i, pheno)
}
\arguments{
\item{i}{list of numeric indices of the data.frame}

\item{pheno}{Data.frame}
}
\description{
Evaluate how many values are empty
}
