% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.R
\name{evaluate_orig}
\alias{evaluate_orig}
\title{Evaluate each variable provided}
\usage{
evaluate_orig(pheno)
}
\arguments{
\item{pheno}{Data}
}
\value{
A matrix with the values
}
\description{
Evaluate each variable provided
}
