% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/own_algorithm.R
\name{evaluate_sd}
\alias{evaluate_sd}
\title{Evaluates the mean of the numeric values}
\usage{
evaluate_sd(i, pheno)
}
\arguments{
\item{i}{List of indices}

\item{pheno}{Data.frame with the samples}
}
\value{
A Matrix with the sd value for each column
}
\description{
Evaluates the mean of the numeric values
}
