% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_num.R
\name{evaluate_sd}
\alias{evaluate_sd}
\title{Evaluates the mean of the numeric values}
\usage{
evaluate_sd(i, pheno)
}
\arguments{
\item{i}{List of indices}

\item{pheno}{Data.frame with the samples}
}
\value{
A matrix with the standard deviation value for each column for each
subset
}
\description{
Looks for the standard deviation of the numeric values
}
