% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experDesign-package.R
\docType{package}
\name{experDesign-package}
\alias{experDesign-package}
\alias{experDesign}
\title{experDesign: Expert experiment design in batches}
\description{
Enables easy distribution of samples per batch avoiding batch and
counfounding effects by randomization of the variables in each batch.
}
\details{
The most important function is \code{\link{design}}, which distributes
samples in batches according to the information provided.

To help in the bench there is the \code{\link{inspect}} function that appends
the group to the data provided.
}
\author{
Lluís Revilla
}
