% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC.R
\name{extreme_cases}
\alias{extreme_cases}
\title{Select the subset of extreme cases to evaluation}
\usage{
extreme_cases(pheno, size, omit = NULL, iterations = 500)
}
\arguments{
\item{pheno}{Data.frame with the sample information.}

\item{size}{The number of samples to subset.}

\item{omit}{Name of the columns of the pheno that will be omitted}

\item{iterations}{Numeric value of iterations that will be performed}
}
\description{
Select the subset of extreme cases to evaluation
}
