% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC.R
\name{extreme_cases}
\alias{extreme_cases}
\title{Select the subset of extreme cases to evaluation}
\usage{
extreme_cases(pheno, size, omit = NULL, each = FALSE,
  iterations = 500)
}
\arguments{
\item{pheno}{Data.frame with the sample information.}

\item{size}{The number of samples to subset.}

\item{omit}{Name of the columns of the pheno that will be omitted}

\item{each}{A logical value if the subset should be taken from all the
samples or for each batch.}

\item{iterations}{Numeric value of iterations that will be performed}
}
\value{
A vector with the number of the rows that are selected.
}
\description{
Subset some samples that are mostly different.
}
\seealso{
\code{\link{optimum_size}}
}
