% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC.R
\name{extreme_cases}
\alias{extreme_cases}
\title{Select the subset of extreme cases to evaluation}
\usage{
extreme_cases(pheno, size, omit = NULL, iterations = 500)
}
\description{
Select the subset of extreme cases to evaluation
}
