% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/follow_up.R
\name{follow_up}
\alias{follow_up}
\title{Follow up experiments}
\usage{
follow_up(original, follow_up, old_new = "batch")
}
\arguments{
\item{original}{A \code{data.frame} with the information of the samples used originally.}

\item{follow_up}{A \code{data.frame} with the information of the new samples.}

\item{old_new}{Name of the column where the batch status will be stored. If
it matches the name of a column in original it will be used to find previous
batches.}
}
\description{
If an experiment was carried out with some samples and you want to continue
with some other samples later on.
}
\seealso{
follow_up2()
}
