% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.R
\name{inspect}
\alias{inspect}
\title{Inspect the index}
\usage{
inspect(i, pheno, omit = NULL)
}
\arguments{
\item{i}{List of indices of samples per batch}

\item{pheno}{Data.frame with the sample information.}

\item{omit}{Name of the columns of the pheno that will be omitted}
}
\value{
The data.frame
}
\description{
Given the index and the data of the samples append the batch assignment
}
