% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC.R
\name{optimum_size}
\alias{optimum_size}
\title{Seek optimum size for a batch}
\usage{
optimum_size(pheno, omit = NULL)
}
\arguments{
\item{pheno}{A \code{data.frame} with the information about the samples.}
}
\value{
A number for the number of samples needed. At most it can be half of
the samples.
}
\description{
Calculates the number of samples that encode all the qualitative variables.
}
