% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/designer.R
\name{replicates}
\alias{replicates}
\title{Design a batch experiment with experimental controls}
\usage{
replicates(pheno, size_subset, controls, omit = NULL, iterations = 500)
}
\arguments{
\item{pheno}{Data.frame with the sample information.}

\item{size_subset}{Numeric value of the number of sample per batch}

\item{controls}{The numeric value of the amount of technical controls per
batch.}

\item{omit}{Name of the columns of the pheno that will be omitted}

\item{iterations}{Numeric value of iterations that will be performed}
}
\value{
A list with to elements the distribution without the controls and
the controls to be replicated.
}
\description{
To ensure that the batches are comparable some samples are processed in each
batch. This function allows to take into account that effect
}
\seealso{
\code{\link{design}}
}
